Refactor: move WebSocket code out of Extension class
